商品削除機能　修正
「削除しようとしているユーザーが出品者かどうか」の条件分岐を追加

+++ app/controllers/items_controller.rb
@@ -38,6 +38,10 @@ class ItemsController < ApplicationController
 
   def destroy
     item = Item.find(params[:id])
+    unless current_user == item.user
+      redirect_to root_path
+      return
+    end
     item.destroy
     redirect_to root_path
   end
